pyproject.toml
[project]
name = "robot-framework-lint-rules"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "robotframework>=7.4.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "robotframework-robocop>=8.2.7",
]


[tool.pre-commit]
exclude = ["TestLibrary.py", "test.resource"]
